# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "distribute API",
    "description": "API Gateway for distribute.\n\n## Quick Start\n\n1. Create an API key in the distribute dashboard, or via `POST /v1/api-keys`\n2. Use it as a Bearer token — that's it, no extra headers needed\n\n```\nAuthorization: Bearer distrib.usr_abc123...\n```\n\nYour key carries your org and user identity. All endpoints work out of the box.\n\n## Storing provider keys (BYOK)\n\nTo store your own provider API keys (e.g. OpenAI, Anthropic) for use in workflows:\n\n```\nPOST /v1/keys\nAuthorization: Bearer distrib.usr_abc123...\n\n{ \"keySource\": \"org\", \"provider\": \"openai\", \"apiKey\": \"sk-...\" }\n```\n\n## Error codes\n\n| Code | Meaning |\n|------|---------|\n| 401 | Missing or invalid Bearer token |\n| 400 | Organization context required (missing `x-org-id` — app key only) |\n| 502 | Identity resolution failed (internal service unreachable) |\n\n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.distribute.you"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication.\n\nUse an API key (`distrib.usr_*`) as your Bearer token. Create one via `POST /v1/api-keys` or in the dashboard.\n\nYour key carries your org and user identity. No extra headers needed."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "BestWorkflowResponse": {
        "type": "object",
        "properties": {
          "workflow": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "category": {
                "type": "string"
              },
              "channel": {
                "type": "string"
              },
              "audienceType": {
                "type": "string"
              },
              "signature": {
                "type": "string"
              },
              "signatureName": {
                "type": "string"
              },
              "humanId": {
                "type": "string",
                "nullable": true,
                "description": "Human ID if styled after an expert"
              },
              "styleName": {
                "type": "string",
                "nullable": true,
                "description": "Base style name for versioning (e.g. 'hormozi')"
              }
            },
            "required": [
              "id",
              "name",
              "category",
              "channel",
              "audienceType",
              "signature",
              "signatureName",
              "humanId",
              "styleName"
            ]
          },
          "dag": {
            "type": "object",
            "properties": {
              "nodes": {
                "type": "array",
                "items": {
                  "nullable": true
                }
              },
              "edges": {
                "type": "array",
                "items": {
                  "nullable": true
                }
              }
            },
            "required": [
              "nodes",
              "edges"
            ]
          },
          "stats": {
            "type": "object",
            "properties": {
              "totalCostInUsdCents": {
                "type": "number"
              },
              "totalOutcomes": {
                "type": "number"
              },
              "costPerOutcome": {
                "type": "number"
              },
              "completedRuns": {
                "type": "number"
              }
            },
            "required": [
              "totalCostInUsdCents",
              "totalOutcomes",
              "costPerOutcome",
              "completedRuns"
            ]
          }
        },
        "required": [
          "workflow",
          "dag",
          "stats"
        ]
      },
...
      "GenerateWorkflowRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "minLength": 10,
            "description": "Natural language description of the desired workflow. Be specific about steps, services, and data flow."
          },
          "hints": {
            "type": "object",
            "properties": {
              "services": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Scope generation to these services"
              },
              "nodeTypes": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Suggest specific node types"
              },
              "expectedInputs": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Expected flow_input field names (e.g. campaignId, email)"
              }
            },
            "description": "Optional hints to guide DAG generation"
          },
          "style": {
            "$ref": "#/components/schemas/WorkflowStyle"
          }
        },
        "required": [
          "description"
        ]
      },
      "WorkflowSummaryResponse": {
        "type": "object",
        "properties": {
          "workflowName": {
            "type": "string",
            "description": "Workflow name"
          },
          "summary": {
            "type": "string",
            "description": "Natural-language summary of the workflow"
          },
          "requiredProviders": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of external provider names required by this workflow"
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Ordered list of workflow steps in human-readable format"
          }
        },
        "required": [
          "workflowName",
          "summary",
          "requiredProviders",
          "steps"
        ]
      },
...
      "WorkflowKeyStatusResponse": {
        "type": "object",
        "properties": {
          "workflowName": {
            "type": "string",
            "description": "Workflow name"
          },
          "ready": {
            "type": "boolean",
            "description": "True if all required provider keys are configured"
          },
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowKeyStatusItem"
            },
            "description": "Status of each required provider key"
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of provider names with missing keys"
          }
        },
        "required": [
          "workflowName",
          "ready",
          "keys",
          "missing"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/workflows": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "List workflows",
        "description": "List available workflows from the workflow-service, optionally filtered by category, channel, or audience type",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Filter by category (e.g. 'sales', 'pr')"
            },
            "required": false,
            "description": "Filter by category (e.g. 'sales', 'pr')",
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by channel (e.g. 'email')"
            },
            "required": false,
            "description": "Filter by channel (e.g. 'email')",
            "name": "channel",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by audience type (e.g. 'cold-outreach')"
            },
            "required": false,
            "description": "Filter by audience type (e.g. 'cold-outreach')",
            "name": "audienceType",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter workflows by human expert ID"
            },
            "required": false,
            "description": "Filter workflows by human expert ID",
            "name": "humanId",
            "in": "query"
          },
          {
            "name": "x-org-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID (e.g. Clerk org ID `org_2xyz...`). Required when using an app key (`distrib.app_*`) on endpoints that need org context. Ignored when using a user key (`distrib.usr_*`)."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External user ID (e.g. Clerk user ID `user_2abc...`). Required when using an app key (`distrib.app_*`) on endpoints that need user context. Ignored when using a user key (`distrib.usr_*`)."
          }
        ],
        "responses": {
          "200": {
            "description": "List of workflows"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{id}": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "Get a workflow",
        "description": "Get a single workflow with full DAG definition",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Workflow ID"
            },
            "required": true,
            "description": "Workflow ID",
            "name": "id",
            "in": "path"
          },
          {
            "name": "x-org-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID (e.g. Clerk org ID `org_2xyz...`). Required when using an app key (`distrib.app_*`) on endpoints that need org context. Ignored when using a user key (`distrib.usr_*`)."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External user ID (e.g. Clerk user ID `user_2abc...`). Required when using an app key (`distrib.app_*`) on endpoints that need user context. Ignored when using a user key (`distrib.usr_*`)."
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow with DAG"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/best": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "Get best-performing workflow",
        "description": "Returns the workflow with the lowest cost per outcome, filtered by category, channel, audience type, and objective",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Filter by category (e.g. 'sales')"
            },
            "required": false,
            "description": "Filter by category (e.g. 'sales')",
            "name": "category",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by channel (e.g. 'email')"
            },
            "required": false,
            "description": "Filter by channel (e.g. 'email')",
            "name": "channel",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by audience type (e.g. 'cold-outreach')"
            },
            "required": false,
            "description": "Filter by audience type (e.g. 'cold-outreach')",
            "name": "audienceType",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Optimization objective ('replies' or 'clicks')"
            },
            "required": false,
            "description": "Optimization objective ('replies' or 'clicks')",
            "name": "objective",
            "in": "query"
          },
          {
            "name": "x-org-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID (e.g. Clerk org ID `org_2xyz...`). Required when using an app key (`distrib.app_*`) on endpoints that need org context. Ignored when using a user key (`distrib.usr_*`)."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External user ID (e.g. Clerk user ID `user_2abc...`). Required when using an app key (`distrib.app_*`) on endpoints that need user context. Ignored when using a user key (`distrib.usr_*`)."
          }
        ],
        "responses": {
          "200": {
            "description": "Best-performing workflow with DAG and stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BestWorkflowResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/generate": {
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Generate a workflow DAG",
        "description": "Uses AI to generate a workflow DAG from a natural language description. The generated workflow is validated and deployed automatically.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateWorkflowRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated and deployed workflow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateWorkflowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Could not generate a valid DAG",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-org-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID (e.g. Clerk org ID `org_2xyz...`). Required when using an app key (`distrib.app_*`) on endpoints that need org context. Ignored when using a user key (`distrib.usr_*`)."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External user ID (e.g. Clerk user ID `user_2abc...`). Required when using an app key (`distrib.app_*`) on endpoints that need user context. Ignored when using a user key (`distrib.usr_*`)."
          }
        ]
      }
    },
    "/v1/workflows/{id}/summary": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "Get workflow summary",
        "description": "Returns a human-readable summary of a workflow's DAG, including ordered steps and required providers. Useful for showing users what a workflow does without exposing the raw DAG.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Workflow ID"
            },
            "required": true,
            "description": "Workflow ID",
            "name": "id",
            "in": "path"
          },
          {
            "name": "x-org-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID (e.g. Clerk org ID `org_2xyz...`). Required when using an app key (`distrib.app_*`) on endpoints that need org context. Ignored when using a user key (`distrib.usr_*`)."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External user ID (e.g. Clerk user ID `user_2abc...`). Required when using an app key (`distrib.app_*`) on endpoints that need user context. Ignored when using a user key (`distrib.usr_*`)."
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSummaryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{id}/key-status": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "Get key status for a workflow",
        "description": "Compares the workflow's required providers against the org's configured BYOK keys. Returns which keys are present and which are missing, along with an overall readiness flag.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Workflow ID"
            },
            "required": true,
            "description": "Workflow ID",
            "name": "id",
            "in": "path"
          },
          {
            "name": "x-org-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External organization ID (e.g. Clerk org ID `org_2xyz...`). Required when using an app key (`distrib.app_*`) on endpoints that need org context. Ignored when using a user key (`distrib.usr_*`)."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "External user ID (e.g. Clerk user ID `user_2abc...`). Required when using an app key (`distrib.app_*`) on endpoints that need user context. Ignored when using a user key (`distrib.usr_*`)."
          }
        ],
        "responses": {
          "200": {
            "description": "Key status for the workflow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowKeyStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}